blacklists.py: load_blacklists() broken out from datahandling
chatcommands.py, datahandling.py, findspam.py:
import load_blacklists from blacklists before use

findspam.py: additionally, actually load blacklists before
defining regexes which depend on them

test/test_spamhandling.py: update import accordingly
