+++ test/test_spamhandling.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 from spamhandling import check_if_spam, check_if_spam_json
-from datahandling import add_blacklisted_user, add_whitelisted_user, load_blacklists
+from datahandling import add_blacklisted_user, add_whitelisted_user
+from blacklists import load_blacklists
 from parsing import get_user_from_url
 import pytest
 import os
